Backend

Documentation

type Ground = Map Int [Block]

The key is the x-coordinate and the value is the blocks at this index

type Indexes = Map Block Int

The key is a block in the world and the value the x-coordinate of this block

data World

Constructors

W

Fields

holding :: Maybe Block

ground :: Ground

indexes :: Indexes

wsize :: Int

findBlockByName :: String -> World -> Maybe Block

Takes the name of a block and returns this block if it exist in the world

isLeftOf :: Block -> Block -> World -> Bool

Check if the first input block is to the left of the second input block

isRightOf :: Block -> Block -> World -> Bool

Check if the first input block is to the right of the second input block

isAbove :: Block -> Block -> World -> Bool

Check if the first input block is above the second input block

isUnder :: Block -> Block -> World -> Bool

Check if the first input block is under the second input block

isOnGround :: Block -> World -> Bool

check if a block is on the ground in a world or equally the block exists in the world and is not the current holding

isOnTop :: Block -> Block -> World -> Bool

Check if a block is on the top of any stack in a world

isOnTop' :: Block -> Block -> World -> Bool

Check if the first input block lies directly over the second input block

isOnBottom :: Block -> World -> Bool

Check if a block is at the bottom of any stack in a world

isBeside :: Block -> Block -> World -> Bool

Check if the first input block is either left or right of secondond input block in a given world

isHolding :: Block -> World -> Bool

Check if a block is a holding in the world

isEmptyIndex :: Int -> World -> Bool

Check if a there exist blocks on a given x-coordinate in world

isOnPoss :: Block -> Int -> World -> Bool

check if a x-coordinate holds the input block

createWorld :: [[String]] -> String -> [Block] -> Maybe World

createGround :: [[String]] -> [Block] -> Maybe Ground

Used by createWorld

createIndexes :: [[String]] -> [Block] -> Maybe Indexes

Used by createWorld

createBlocks :: Int -> [String] -> [Block] -> Maybe [(Block, Int)]

Used by createWorld

createBlocks' :: Int -> [String] -> [Block] -> Maybe (Int, [Block])

Used by createWorld

getBlock :: String -> [Block] -> Maybe Block

Input : a name of a block , any list of blocks Output: The block with the given name from the list if it exist

getBlockIndex :: Block -> World -> Maybe Int

Input: Any block, a word return x-coordinate of the given block if it exist in the world

getBlocksAt :: Int -> World -> Maybe [Block]

Input: A x-coordinate, a world Output: All blocks at the given index if the index exists in the world

getBlocksOnGroundBy :: (Block -> Bool) -> World -> [Block]

Input: a function, a world Output Returns all blocks an the ground that satifies the given function

getRightMost :: [Block] -> World -> Maybe Block

Input: Any list a blocks, a world Output: The first occurance of a block with the biggest x-coordinate if all given blocks exist on the ground in the given list

getLeftMost :: [Block] -> World -> Maybe Block

The same as rightmost but with smallest index

getUnderMost :: [Block] -> World -> Maybe Block

Input: any list of blocks, a world Output: If the input blocks havn't the same index nothing is returned else the bottom block from the input list in the stack

getUpperMost :: [Block] -> World -> Maybe Block

getUpperUnderMost :: ([Block] -> Block) -> [Block] -> World -> Maybe Block

Used by getUnderMost and getUpperMost

initWorld :: [[[Char]]]

For testing purposes