NLPParser

Documentation

data Action

Constructors

Move

Put

Take

None

data Reference

Constructors

Above
Mustn't be directly above a block.

Beside
Either left of or right of

Inside

LeftOf

OnTop
Uppermost possition i.e. ontop of a stack

RightOf

Under
Mustn't be directly above a block.

data Location

Constructors

Empty

Location Reference [Block]

Floor [Int]

data Output

Constructors

O

Fields

action :: Action

An action to take.

mBlocks :: [Block]

Blocks to move.

location :: Location

An location to put the blocks at.

initOutput :: Output

command :: String

modifyString :: String -> String

tmpMain :: IO ()

For testing purposes

runParser :: PGF -> String -> World -> [Err Output]

traverseTree :: GS -> World -> Err Output

there can be copies of the same block in the final mBlocks, this is expected and solved by nub

handleThing :: GThing -> World -> Err [Block]

TODO think about the differance between the and all

handleLocation :: GLocation -> [Block] -> World -> Err [Block]

filterBlocks :: (Block -> Block -> World -> Bool) -> World -> GThing -> [Block] -> Err [Block]

pairBlocks :: [Block] -> [Block] -> [(Block, Block)]

getRefLocation :: GLocation -> World -> Err Location

TODO it can be possbile to put something above all blocks depending on the world

handleGBlock :: GBlock -> World -> Err [Block]

handleBlock :: GBlock -> World -> [Block]

world :: [[String]]

For debugging purposes